requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==21.2.0
autopage==0.4.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.5.0
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cliff==3.10.0
cmd2==0.0.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debtcollector==2.3.0
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.1.4
django-jazzmin==3.0.1
django-taggit==6.1.0
djangorestframework==3.15.2
docutils==0.17.1
drf-yasg==1.21.8
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
frozenlist==1.5.0
future==0.18.2
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iso8601==1.0.2
jeepney==0.7.1
keyring==23.5.0
keystoneauth1==4.4.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libusb1==3.2.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.1.0
ndeflib==0.3.3
netaddr==0.8.0
netifaces==0.11.0
nfcpy==1.0.4
oauthlib==3.2.0
olefile==0.46
os-service-types==1.7.0
oslo.i18n==5.1.0
oslo.serialization==4.2.0
oslo.utils==4.12.2
packaging==21.3
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
prettytable==2.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pyDes==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
pyscard==2.2.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-freezerclient==4.3.0
python-telegram-bot==21.9
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rfc3986==1.5.0
roman==3.3
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
sqlparse==0.4.2
stevedore==3.5.0
systemd-python==234
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
